[Exposing ALTS Context 2/2] Utility Wrapper Class
This PR serves as the same purpose as this reverted one:

https://github.com/grpc/grpc/pull/21196,

but we adopt another approach for providing utilities to get AltsContext.
In that PR, we use upb to serialize the AltsContext in TSI layer, pass it up to the
C++ user-facing APIs, and use standard pb to de-serialize it. Here we
use both upb for serializing and de-serializing, and provide a utility
wrapper class for  the class generated from upb.
This change is based on the assumption that upb generated class is not
user-friendly enough and we'd better provide more straightforward ways
for users to use ALTS context.